+++ jdroid-gradle-android-plugin/src/main/groovy/com/jdroid/gradle/android/AndroidGradlePlugin.groovy
@@ -9,13 +9,13 @@ import org.gradle.api.Project
 
 public abstract class AndroidGradlePlugin extends JavaBaseGradlePlugin {
 
-	private static final int ANDROID_SDK_VERSION = 27
+	private static final int ANDROID_SDK_VERSION = 28
 
 	// http://developer.android.com/tools/revisions/build-tools.html
-	private static final String ANDROID_BUILD_TOOLS_VERSION = '27.0.3'
+	private static final String ANDROID_BUILD_TOOLS_VERSION = '28.0.2'
 
 	// https://github.com/maxirosson/jdroid-android-lint/releases
-	private static final String JDROID_ANDROID_LINT_RULES_VERSION = "1.0.0"
+	private static final String JDROID_ANDROID_LINT_RULES_VERSION = "1.1.0"
 
 	protected BaseExtension android
 
@@ -40,12 +40,12 @@ public abstract class AndroidGradlePlugin extends JavaBaseGradlePlugin {
 			}
 		}
 
-		android.compileSdkVersion propertyResolver.getIntegerProp('ANDROID_COMPILE_SDK_VERSION', ANDROID_SDK_VERSION)
-		android.buildToolsVersion propertyResolver.getStringProp('ANDROID_BUILD_TOOLS_VERSION', ANDROID_BUILD_TOOLS_VERSION)
+		android.setCompileSdkVersion(propertyResolver.getIntegerProp('ANDROID_COMPILE_SDK_VERSION', ANDROID_SDK_VERSION));
+		android.setBuildToolsVersion(propertyResolver.getStringProp('ANDROID_BUILD_TOOLS_VERSION', ANDROID_BUILD_TOOLS_VERSION));
+		android.getDefaultConfig().setMinSdkVersion(jdroid.minimumSdkVersion);
+		android.getDefaultConfig().setTargetSdkVersion(propertyResolver.getIntegerProp('ANDROID_TARGET_SDK_VERSION', ANDROID_SDK_VERSION));
 
 		android.defaultConfig {
-			minSdkVersion jdroid.minimumSdkVersion
-			targetSdkVersion propertyResolver.getIntegerProp('ANDROID_TARGET_SDK_VERSION', ANDROID_SDK_VERSION)
 
 			vectorDrawables.useSupportLibrary = propertyResolver.getBooleanProp('VECTOR_DRAWABLES_USE_SUPPORT_LIB', true)
 
@@ -83,7 +83,7 @@ public abstract class AndroidGradlePlugin extends JavaBaseGradlePlugin {
 			abortOnError propertyResolver.getBooleanProp('ABORT_ON_LINT_ERROR', true)
 			enable 'ConvertToWebp'
 			disable 'ContentDescription', 'RtlEnabled', 'RtlHardcoded', 'RtlSymmetry', 'UseCompoundDrawables', 'UnknownIdInLayout', 'RequiredSize'
-			error 'StringFormatMatches', 'StringFormatCount', 'HardcodedText', 'ScrollViewSize', 'SwitchIntDef', 'Deprecated', 'TextFields'
+			error 'StringFormatMatches', 'StringFormatCount', 'HardcodedText', 'ScrollViewSize', 'SwitchIntDef', 'Deprecated', 'TextFields', 'GradleDynamicVersion', 'DefaultLocale'
 		}
 
 		android.packagingOptions {