+++ jdroid-gradle-project-plugin/src/main/groovy/com/jdroid/gradle/GradleProjectPlugin.groovy
@@ -11,10 +11,10 @@ public class GradleProjectPlugin extends JavaBaseGradlePlugin {
 
 		project.ext.PACKAGING = 'jar'
 
-		project.apply plugin: 'groovy'
+		applyPlugin("groovy");
 
 		// https://docs.gradle.org/current/userguide/javaGradle_plugin.html
-		project.apply plugin: 'java-gradle-plugin'
+		applyPlugin("java-gradle-plugin");
 
 		project.dependencies {
 			compile localGroovy()